Allow users to specifiy the encoding of ZPL code in socket messages
Windows still largely uses Latin 1 (aka. ISO 8859-1) to encode data. While
the first blocks of Latin 1 are identical to UTF-8, trying to decode
Latin 1 encoded texts will start to cause issues as soon as special
characters (like the German umlauts) are being used.

In order to avoid fixing the encoding to a single character set, this
commit introduces the ability to select the encoding individually in
the global program settings.

Only UTF 8 and Latin 1 are offered as available choices so far, but every
encoding supported by .NET could be easily added to the selection. The
entire list of encodings can be found at
https://learn.microsoft.com/en-us/dotnet/api/system.text.encoding?view=net-7.0#list-of-encodings